+++ .travis.yml
@@ -8,7 +8,7 @@ install:
     - pip install coveralls
 
 script:
-    python runtests.py --with-coverage  --cover-package=package,common,models --exclude-dir=integration
+    python runtests.py --with-coverage  --cover-package=package --exclude-dir=integration
 
 after_success:
     coveralls
