+++ integration/integration_commands/integration_test_refresh_ip.py
@@ -6,8 +6,8 @@ from pyVim.connect import SmartConnect, Disconnect
 from cloudshell.cp.vcenter.commands.refresh_ip import RefreshIpCommand
 from cloudshell.cp.vcenter.common.vcenter.vmomi_service import pyVmomiService
 from cloudshell.cp.vcenter.models.VCenterConnectionDetails import VCenterConnectionDetails
-from tests.utils import helpers
-from tests.utils.testing_credentials import TestCredentials
+from cloudshell.tests.utils import TestCredentials
+from cloudshell.tests.utils import helpers
 
 
 class TestRefreshIpCommand(TestCase):
