+++ package/cloudshell/tests/test_commands/test_connect_vm.py
