+++ package/cloudshell/tests/test_commands/test_deploy_vm.py
