+++ package/cloudshell/tests/test_common/test_cloudshell/test_connection_detail_retriever.py
