+++ package/cloudshell/tests/test_common/test_vcenter/test_ovf_image_tool.py
