+++ package/cloudshell/tests/test_common/test_vcenter/test_vcenter_data_model_retriever.py
