+++ package/cloudshell/tests/test_common/test_vcenter/test_vm_location.py
