+++ package/cloudshell/tests/test_models/test_deploy_data_holder.py
