+++ package/cloudshell/tests/test_network/test_dvswitch/test_name_generator.py
