+++ package/cloudshell/tests/test_network/vnic/test_network_vnic.py
