+++ package/cloudshell/tests/test_vm/test_vnic_to_network_mapper.py
