+++ runtests.bat
@@ -1,0 +1,3 @@
+cd package
+python setup.py test
+cd /d %~dp0
