+++ runtests.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+cd package
+python setup.py test
